Refactorizacion de  token
He tenido que mover la funcion para pillar el token a la pagina de inicio porque su useEffect se ejecuta antes que el de App y de paso he puesto que cada vez que se reinicie el objeto spotifyApiObject ya pille el token automaticamente alli
