+++ Dutify/src/App.jsx
@@ -31,15 +31,7 @@ function App() {
   const [open, setOpen] = useState(false);
 
   useEffect(() => {
-    let spotifyToken = window.sessionStorage.getItem("token");
-
-    if (!spotifyToken || spotifyToken === "undefined") {
-      spotifyToken = getTokenFromUrl().access_token;
-      window.sessionStorage.setItem("token", spotifyToken);
-    }
-
-    setToken(spotifyToken);
-    setAccessToken(spotifyToken);
+    console.log("useEffect de App.jsx")
 
     const currentTrack = JSON.parse(window.sessionStorage.getItem("currentTrack"));
     const currentTime = window.sessionStorage.getItem("currentTrackTime")
@@ -52,6 +44,32 @@ function App() {
     } 
   }, []);
 
+  const AUTH_ENDPOINT = "https://accounts.spotify.com/authorize";
+  const REDIRECT_URI = "http://localhost:5173/inicio";
+  const CLIENT_ID = "212f24bfe4124f9d89ee2c341ae96f19";
+  const RESPONSE_TYPE = "token";
+
+  const scopes = [
+    "user-read-currently-playing",
+    "user-read-playback-state",
+    "user-modify-playback-state",
+    "app-remote-control",
+    "playlist-modify-public",
+    "playlist-modify-private",
+    "playlist-read-private",
+    "playlist-read-collaborative",
+    "user-top-read",
+    "user-library-read",
+    "streaming",
+    "user-read-email",
+    "user-read-private",
+    "user-modify-playback-state",
+    "user-read-playback-state"
+  ];
+
+  const loginUrl = `${AUTH_ENDPOINT}?client_id=${CLIENT_ID}&redirect_uri=${REDIRECT_URI}&response_type=${RESPONSE_TYPE}&scope=${scopes.join(
+    "%20"
+  )}&show_dialog=true`;
 
   const router = createBrowserRouter([
     {
@@ -88,44 +106,6 @@ function App() {
     },
   ]);
 
-  const getTokenFromUrl = () => {
-    return window.location.hash
-      .substring(1)
-      .split("&")
-      .reduce((initial, item) => {
-        let parts = item.split("=");
-        initial[parts[0]] = decodeURIComponent(parts[1]);
-        return initial;
-      }, {});
-  };
-
-  const AUTH_ENDPOINT = "https://accounts.spotify.com/authorize";
-  const REDIRECT_URI = "http://localhost:5173/";
-  const CLIENT_ID = "212f24bfe4124f9d89ee2c341ae96f19";
-  const RESPONSE_TYPE = "token";
-
-  const scopes = [
-    "user-read-currently-playing",
-    "user-read-playback-state",
-    "user-modify-playback-state",
-    "app-remote-control",
-    "playlist-modify-public",
-    "playlist-modify-private",
-    "playlist-read-private",
-    "playlist-read-collaborative",
-    "user-top-read",
-    "user-library-read",
-    "streaming",
-    "user-read-email",
-    "user-read-private",
-    "user-modify-playback-state",
-    "user-read-playback-state"
-  ];
-
-  const loginUrl = `${AUTH_ENDPOINT}?client_id=${CLIENT_ID}&redirect_uri=${REDIRECT_URI}&response_type=${RESPONSE_TYPE}&scope=${scopes.join(
-    "%20"
-  )}&show_dialog=true`;
-
   const handleClose = () => {
     setFeedback("");
     setOpen(false);
@@ -155,7 +135,6 @@ function App() {
                   </ClickAwayListener>
               ) : null}
             </div>
-
             <TopBar></TopBar>
             <main>
               <RouterProvider router={router}></RouterProvider>
