+++ Dutify/src/components/locations/login/login.jsx
@@ -9,7 +9,7 @@ function Login(loginUrl) {
     <section className="login-section d-flex flex-column align-items-center justify-content-center h-100 w-100 gap-5">
       <div className="d-flex flex-column align-items-center justify-content-center">
         <h1 className="title">Dutify</h1>
-        <BsSpotify size={200}></BsSpotify>
+        <BsSpotify size={220}></BsSpotify>
       </div>
       <button className="btn-login" onClick={goToLoginUrl}>Iniciar Sesion con Spotify</button>
     </section>
